fix: resolve clippy warnings in ftui-render, ftui-layout, ftui-widgets
- ftui-render/diff.rs: remove unnecessary usize->usize cast
- ftui-layout/lib.rs: remove unused last_weighted_pos variable and
  prefix unused cache/id destructured bindings with underscore
- ftui-widgets/textarea.rs: remove redundant usize casts on
  last_viewport_height.get()

+++ crates/ftui-render/src/diff.rs
@@ -1810,7 +1810,7 @@ mod tests {
         }
 
         // Invariant 2: total cells in runs == diff.len()
-        let total_cells: usize = runs.iter().map(|r| r.len() as usize).sum();
+        let total_cells: usize = runs.iter().map(|r| r.len()).sum();
         assert_eq!(
             total_cells,
             diff.len(),
